requirements.txt
# Requirements for the Python project
# Add your project dependencies here.
pytest
streamlit>=1.32.0
mistralai>=0.0.10
python-dotenv>=1.0.0
streamlit_pdf_viewer